refactor(jwt): custom devise strategy combined with access list
There are now two types of jwt Tokens used in the App. Short Term Tokens which expire after some hours
for the use in the angular frontend for example and Tokens without or with longer expiration times for use in
API clients, Mobile Apps etc. The latter includes an auth_token which is generated and saved in a db table with
the used user agent. This allows the User to see which user agents are accessing his data and can remove access
to a device for example.